refactor(core): use Polyfill package
Closes #1329
