--- src/Dekaf/Compatibility/Net10ValueTaskForwarders.cs
@@ -1,17 +1,0 @@
-#if !NETSTANDARD2_0
-namespace System.Threading.Tasks;
-
-internal static class ValueTaskCompatibility
-{
-    public static ValueTask CompletedTask => ValueTask.CompletedTask;
-
-    public static ValueTask FromCanceled(System.Threading.CancellationToken cancellationToken)
-        => ValueTask.FromCanceled(cancellationToken);
-
-    public static ValueTask FromException(Exception exception)
-        => ValueTask.FromException(exception);
-
-    public static ValueTask<T> FromException<T>(Exception exception)
-        => ValueTask.FromException<T>(exception);
-}
-#endif
