+++ src/Dekaf/Dekaf.csproj
@@ -31,10 +31,12 @@
   <PropertyGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
     <NoWarn>$(NoWarn);NETSDK1138</NoWarn>
     <PublishAot>false</PublishAot>
+    <PolyArgumentExceptions>true</PolyArgumentExceptions>
   </PropertyGroup>
 
   <!-- Compatibility dependencies for the package's netstandard2.0 asset. -->
   <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+    <PackageReference Include="Polyfill" Version="10.11.2" PrivateAssets="all" />
     <PackageReference Include="System.IO.Pipelines" Version="10.0.9" />
     <PackageReference Include="System.Text.Json" Version="10.0.9" />
     <PackageReference Include="System.Threading.Channels" Version="10.0.9" />
