+++ src/Dekaf/Producer/ValueTaskSourcePool.cs
@@ -137,10 +137,10 @@ public sealed class ValueTaskSourcePool<T> : IAsyncDisposable
     public ValueTask DisposeAsync()
     {
         if (Interlocked.Exchange(ref _disposed, 1) != 0)
-            return ValueTaskCompatibility.CompletedTask;
+            return default;
 
         _stack.Clear();
 
-        return ValueTaskCompatibility.CompletedTask;
+        return default;
     }
 }
