+++ src/Dekaf/Protocol/KafkaProtocolReader.cs
@@ -302,7 +302,7 @@ public ref struct KafkaProtocolReader
         {
             if (_position + 16 > _span.Length)
                 ThrowInsufficientData();
-            var result = Dekaf.Compatibility.GuidCompatibility.ReadBigEndian(_span.Slice(_position, 16));
+            var result = Dekaf.KafkaGuid.ReadBigEndian(_span.Slice(_position, 16));
             _position += 16;
             return result;
         }
@@ -314,7 +314,7 @@ public ref struct KafkaProtocolReader
     {
         if (_reader.UnreadSpan.Length >= 16)
         {
-            var result = Dekaf.Compatibility.GuidCompatibility.ReadBigEndian(_reader.UnreadSpan[..16]);
+            var result = Dekaf.KafkaGuid.ReadBigEndian(_reader.UnreadSpan[..16]);
             _reader.Advance(16);
             return result;
         }
@@ -322,7 +322,7 @@ public ref struct KafkaProtocolReader
         if (!_reader.TryCopyTo(buffer))
             ThrowInsufficientData();
         _reader.Advance(16);
-        return Dekaf.Compatibility.GuidCompatibility.ReadBigEndian(buffer);
+        return Dekaf.KafkaGuid.ReadBigEndian(buffer);
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
